requirements.txt
pymavlink
pynput
pyserial
opencv-python
numpy
matplotlib
tqdm
pyyaml
Pillow
